chore: separate matrix and added base class for agents, allowed to pass in the args to tool
